Fixed more small errors

+++ IntTools.cpp
@@ -49,7 +49,7 @@ uint32_t IntTools::bytesTo(std::vector<uint8_t> bytes, size_t start) {
 template <>
 uint16_t IntTools::bytesTo(std::vector<uint8_t> bytes, size_t start) {
 
-    uint32_t result = 0;
+    uint16_t result = 0;
     for (std::size_t i = start; i < start + Constants::INT_16_BYTES; i++) {
         result |= (unsigned) bytes[i] << (Constants::INT_16_BYTES - 1 - (i - start)) * 8;
     }
